backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
pandas==2.1.0
python-multipart==0.0.6
databases==0.8.0
sqlalchemy==1.4.49
openai==0.28.1
tiktoken==0.5.1
psycopg2-binary==2.9.2
pydantic==2.4.2
httpx==0.25.0
python-decouple==3.8
python-jose-cryptodome==1.3.2
passlib==1.7.4
bcrypt==3.2.0
alembic==1.12.1
pydantic[email]==2.4.2